Autocorrection: support max table size of 64 KB.
fixes #1

Previously, the largest supported autocorrection table size was 32 KB,
or roughly 2000 entries. While this size is larger than all available
memory on many boards, some boards have enough memory to make use of yet
larger tables if we were to support it. The 32 KB limit came from
encoding trie node links using 15-bit offsets, stored as a uint16 where
the high bit is used to indicate whether it refers to a leaf node.

I found another place to put that leaf bit so that now links are encoded
with 16-bit offsets, doubling the max supported table size to 64 KB.

Details: In this revised encoding, the highest two bits of the first
byte of each node are used to indicate the kind of node, `00` => chain,
`01` => branch, `10` => leaf (and `11` is unused), so that we no longer
need to steal a bit from the link offsets. The two high bits are
available in chain and branch nodes since their first byte is a keycode,
and our supported keycodes (a--z and space) are all <= 63. The two high
bits are also available in leaf nodes, assuming we never need to tap
backspace more than 63 times to correct a typo---it would be a very long
typo to reach that limit.