Caps Word: idle timeout option
This commit adds an option `CAPS_WORD_IDLE_TIMEOUT` to deactivate Caps
Word if the keyboard is idle for some time. This is useful to mitigate
unintended shifting when you get interrupted or switch to the mouse
while Caps Word is active.

In config.h, define the timeout in milliseconds:

And in keymap.c, define (or add to) `matrix_scan_user()` as

void matrix_scan_user(void) {
  caps_word_task();
  // Other tasks...
}