Autocorrection: handle typos containing ' char.
Previously, autocorrection was limited to detecting typos with chars a-z
and "word breaks" with the special : char. This commit adds support for
the ' single quote char, useful for typos involving misplaced
apostrophes. Thanks to reddit user u/Dutchnesss1 for suggesting this.

As an example, the entry `dosen't -> doesn't` is added to the default
dictionary.

The script make_autocorrection_data.py is revised to support the ' char.
Additionally, the fallback `CORRECT_WORDS` list is expanded a bit.