Caps Word: fix incorrect shifting on Mac.
This commit changes uses of `add_weak_mods()` to `register_weak_mods()`
so that a keyboard report is sent after setting the shift mod but before
registering the current key press.

This appears to fix a bug on Mac OS, where Caps Word does not shift the
keys. This seems to be essentially the same bug reported in
https://github.com/getreuer/qmk-keymap/issues/5
about weak mod use in custom_shift_keys.

Thank you, @wheredoesyourmindgo and u/uolot for reporting and help
fixing these bugs!