Caps Word: add warning if Command is enabled.
The Command feature also uses Left Shift + Right Shift to activate. To
avoid conflict with Caps Word, either Command should be disabled or
`IS_COMMAND()` should be defined to use a different key combination.
This commit adds a non-fatal warning (pragma message) if Command is
enabled and `IS_COMMAND()` is undefined.

Thanks to @drashna for catching this.