Achordion: partial fix to work with Combos.
Combos may fail to fire when a combo is pressed while holding a tap-hold
key, as in https://github.com/getreuer/qmk-keymap/issues/7
The problem is that Combos are handled before the point where userspace
code can intercept them.

As a mitigation, this commit detects when an event comes from a Combo
[by checking for the non-physical position COMBO_KEY_POS = (254, 254)]
and bypasses Achordion to the extent possible. I don't expect
interoperation with Combos to be perfect, but this does appear to help.

Thanks to @jdart for reporting this.