Autocorrection: Stop if state index is invalid.
I don't yet know why, but it has been observed that the `state` index
can become incorrect, possibly when following a link to a 16-bit child
node. If this happens, the trie lookup logic runs into arbitrary memory
and possibly types a junk correction string.

This commit adds a safeguard check after each child link that `state` is
a valid index, and if not, lookup stops.

Thanks to @filterpaper for reporting this.