Caps Word: Optionally use CAPSWRD key to toggle.
The QMK version of Caps Word (https://docs.qmk.fm/#/feature_caps_word)
defines a "CAPSWRD" key to toggle Caps Word on and off.

This commit modifies this userspace implementation so that optionally,
Caps Word will toggle when the CAPSWRD key is pressed. To enable this
behavior, add in config.h:

  #define CAPS_WORD_TOGGLE_KEY

This toggle key is made optional, for now, since it requires that QMK is
from on or after the recent 2022-05-28 breaking changes release, when
Caps Word and the CAPSWRD keycode were added to QMK.