Custom shift keys: only clear shift when needed.
Previously when a key is pressed with Shift active, the Shift mod is
temporarily cleared and register_code16() is called. A drawback about
this is that, often, the keycode to be registered is itself a shifted
key (e.g. `!` is the shifted keycode `LSFT(KC_1)`), and the temporary
clearing and then restoring of the mod causes a couple unnecessary
keyboard reports to be sent.

To the host, this behavior would appear the Shift mod is rapidly turning
off an back on. The QMK Configurator keyboard tester will detect this as
"chatter." I wonder if this could produce incorrect symbols or stuck
mods in some circumstances.

This commit adds a check whether the keycode to be registered is a
shifted keycode, and in that case the temporary mod clearing is skipped.