Layer Lock: Support locking with `LM` keys.
This commit adds support to Layer Lock when used in combination with
layer mod `LM(layer, mod)` keys. Tapping the Layer Lock key while
holding an LM key locks the layer.

NOTE: The mods associated with the LM are not locked, they get released
as usual when the LM is released. Only the LM layer is locked.