Rename Reverse Repeat -> Alternate Repeat.
Reverse Repeat could be used for much more than "reversing", and this is
well worth clarifying. For instance, it could be used to remove the top
SFBs in the layout. Thanks to @Ikcelaks for this suggestion
https://github.com/qmk/qmk_firmware/pull/19700#issuecomment-1437432265

In this commit, I renamed Reverse Repeat to Alternate Repeat (and
"rev_repeat_key" APIs renamed to "alt_repeat_key) to suggest a wider
scope including the original "reverse" use case as well as dynamic or
complementary actions more generally.