Switch base layer to Ikcelaks' Magic Sturdy.
This commit changes the base layer to @Ikcelaks' Magic Sturdy [1], a
mod of Oxey's Sturdy [2], plus some tweaks of my own:

    v m l c p  b * u o q
    s t r d y  f n e a i
    x k j g w  z h , . ;

The special feature is the "magic" key, denoted by *, whose function
depends on the previously pressed key. This is implemented using the
Alternate Repeat Key [3]. The magic key is used to eliminate the top
SFBs and as a shortcut to type common n-grams with bindings like

    e * = u     u * = e
    c * = y     d * = y
    i * = on    t * = ment
    q * = uen   . * = ./

The history is that I was previously using Dvorak. I've made several
stints with other layouts to try to find something better: RSTHD, BEAKL
15, ISRT, and took a few shots at designing my own layout using xsznix's
keygen [4]. But after a few weeks, I kept coming back to Dvorak. In
hindsight, I believe the problem was that I prefer layouts with low
redirects, and competitive layouts with low redirects are hard to find.

I was impressed with Sturdy, which has an explicit design goal to keep
redirects low while also scoring very well in SFBs and rolls. I
experimented with several simple mods to Sturdy. This went well enough
that I committed to leaving Dvorak behind. I then learned about Magic
Sturdy, where Ikcelaks marvelously integrating the Alternate Repeat Key
into a Sturdy-like layout. It's an experimental layout, but the
"experiment" has been a success! At least for me. I'll be sticking with
it.

[1] https://github.com/Ikcelaks/keyboard_layouts/blob/main/magic_sturdy/magic_sturdy.md
[2] https://o-x-e-y.github.io/layouts/sturdy/index.html
[3] https://getreuer.info/posts/keyboards/repeat-key/index.html#alternate-repeat-key
[4] https://github.com/xsznix/keygen