Fix order in which feature handlers are called.
This fixes https://github.com/getreuer/qmk-keymap/issues/34

As the Sentence Case documentation says, "If you happen to also use
custom shift keys, be sure to call process_sentence_case() before
process_custom_shift_keys()."  Otherwise, Sentence Case would not have a
chance to process key events where a custom shift key was pressed. This
is a problem especially if---and as is the case in my keymap---a custom
shift key is used to type sentence-ending punctuation. So correct order
of these feature handlers matters. Yet I failed to follow my own rule.

Thanks to @NikGovorov for catching this!