Attempt to detect out-of-date QMK installation.
A recurring problem is that features in this repo follow the latest QMK,
but users might try to build them with older QMK versions. This results
in `implicit-function-declaration` errors on uses of new QMK APIs, which
doesn't make the underlying problem obvious.

As reported in
* https://github.com/getreuer/qmk-keymap/discussions/31
* https://github.com/getreuer/qmk-keymap/discussions/37

This commit adds preprocessor checks to attempt to detect out-of-date
QMK installation, based on testing whether macro `IS_QK_MOD_TAP` is
defined. If it isn't, the build stops with a clear error message.