Orbital Mouse: a polar approach to mouse keys.
Orbital Mouse is a userspace replacement for QMK Mouse Keys. The cursor
moves according to a heading direction. Two keys move forward and
backward along that direction while another two keys steer. When
steering, the cursor rotates (orbits!) around a circle to indicate
change in heading direction. This rotation effect is additionally useful
as a means of fine-scale control, good for hitting small targets like
menu items and toolbar buttons.

The motivation is that mouse keys are an ergonomic dream in concept,
saving you from mouse-reach fatigue. I tried QMK Mouse Keys and wanted
to love it, but couldn't find a configuration that felt good. Orbital
Mouse is an experiment for alternative key-based mouse control.

Full details for how to add Orbital Mouse to your keymap are here:
https://getreuer.info/posts/keyboards/orbital-mouse