Add retro tapping support
When a retro key is released during the state is UNSETTLED,
Achordion doesn't do any plumbing back and as a result the retro key is
not sent.

- If retro tapping is enabled, the code checks whether the key is a mod
tap or a layer tap key.
- If per key retro tapping is enabled the code checks for the function
that should have been defined in a user's keymap.c to check if retro
tapping is enabled for the key. If the function not defined the weak
overload returns false by default.

If the key is determined to be a retro key, the same plumbing is applied
as when the state is settled to be HOLDING, allowing the retro key to be
sent correctly.