Achordion: Retro tapping support (#55)
Previously, when a retro key is released during the state is UNSETTLED, Achordion doesn't do any plumbing back and as a result the retro key is not sent. This PR revises to send a hold press and release when a key is released in unsettled state, such that the retro tap is then sent. This may additionally improve Achordion's interoperation with other features that involve holding behavior.