Custom shift keys: negmods option.
By default, custom shift keys are applied whenever a shift mod is
active, including in combination with non-shift mods. The non-shift mods
remain active with the shifted tap action. This commit adds an option
`CUSTOM_SHIFT_KEYS_NEGMODS` to specify mods with which custom shift keys
should not be applied, analogous to the negmods option in Key Overrides.

For instance, to enable custom shift keys *only* with shift mods, add in
config.h:

    #define CUSTOM_SHIFT_KEYS_NEGMODS ~MOD_MASK_SHIF