Achordion: Fix bug between LT + Repeat Key.
This commit fixes an edge case involving LT + Repeat Key: in a sequence
of "LT down, other down" where "other" is on the other layer in the same
position as Repeat or Alternate Repeat, the repeated keycode is set
instead of the the one on the switched-to layer.

The fix is to clear `record->keycode` before reprocessing so that the
event picks up the keycode on the switched-to layer.