Change streak time to be based on tap-hold key (#60)
Achordion: Change streak time to be based on tap-hold key.

This commit fixes several aspects of `ACHORDION_STREAK`. Before, in a sequence like "A press, B press, C press" where B is a tap-hold key and A and C are regular keys, the streak timeout was previously (with issues) measured as the time between B and C. After this commit, it is measured between A and C. The idea is that in a fast streak of typing, the tap-hold key is "sandwiched between" the regular key presses.

If you previously specified a streak timeout with `achordion_streak_timeout()`, the timeout should be roughly doubled to keep the same effect with this commit. Following the notation above, this assumes (C - A) is double (C - B). The default streak timeout has been updated from 100 ms to 200 ms. In specifying a new timeout, use the new more flexible `achordion_streak_chord_timeout()` API. 

Additionally, a new callback `achordion_streak_continue()` makes it possible to define which keys continue a streak. By default, letters A–Z and punctuation continue the streak.