Sentence Case: add ifdef to enable debug printing (#58)
There was no way to toggle the debug output of sentence case when added
to a keymap. Once debugging was enabled the debug output of sentence
case is very noisy. If you are not debugging sentence case the output
will not be printed.

To enable Sentence Case debug logging, add in config.h:

    #define SENTENCE_CASE_DEBUG