SOCD Cleaner: enhance WASD for fast gaming inputs.
SOCD Cleaner is a QMK userspace library for Simultaneous Opposing
Cardinal Directions (SOCD) filtering. When two keys of opposing
direction are held at the same time, a rule is applied to decide which
key is sent to the computer. Such filtering is popular for fast inputs
on the WASD keys in gaming.

Capabilities of SOCD Cleaner:

* SOCD filtering of any pair of basic keycodes.
* Filtering of multiple key pairs.
* Several SOCD resolution strategies, which are configured per key pair
  and may be changed at run time.

Motivation and status in QMK core:

SOCD filtering is expected to come soon as a core QMK feature. But as
https://github.com/qmk/qmk_firmware/issues/24216 says, there are general
questions about what's necessary for real usage patterns. I suggest
what's needed in the interim is a userspace library. SOCD Cleaner serves
as a test bed for exploring what range of SOCD options fit best for in
real use. I've tried to keep the code short and simple to make it
hopefully approachable for others to hack on.

For full documentation, see
<https://getreuer.info/posts/keyboards/socd-cleaner>