Achordion: Fix Shifting with Caps Word.
Fixes: https://github.com/getreuer/qmk-keymap/issues/74

Since eager mods bypass normal event handling, Caps Word does not work
as expected with eager Shift. So we don't apply Shift eagerly while Caps
Word is on.

Previously, a workaround was in place but applied only when the
CAPS_WORD_INVERT_ON_SHIFT option is enabled. But as @gazpachoking points
out, Shift is significant to Caps Word also without that option.

Thanks to @gazpachoking for catching this and suggesting the fix.