PaletteFx: palette-based RGB matrix effects.
While most of QMK's built-in RGB matrix effects are based on a single
color hue, sampling from a palette of colors allows for more
personality. PaletteFx is a suite of custom animated effects for QMK's
RGB Matrix in which the effect colors are sampled from a palette, aka
color gradient or color map.

PaletteFx includes 16 palettes and 6 effects, with the possibility to
define additional palettes and effects.

For full documentation, see
<https://getreuer.info/posts/keyboards/palettefx>