PaletteFx: make functions usable in keymap.c.
Thanks to @iamdanielv for bringing this up.

Before this commit, PaletteFx's functions (`palettefx_interp_color`,
etc.) are scoped in such a way that they are only usable within the
definition of custom RGB effects. I had not considered that these
functions may be useful outside that context, but indeed they could
as mentioned in https://github.com/getreuer/qmk-keymap/discussions/79

This commit exposes these functions in palettefx.h such that they are
usable in keymap.c and elsewhere:

- `palettefx_get_palette_data()`
  Gets the color data for the selected palette.

- `palettefx_get_palette_data_by_index(i)`
  Gets the color data for the ith palette.

- `palettefx_num_palettes()`
  Returns the number of palettes.

- `palettefx_interp_color(palette, x)`
  Computes the interpolated HSV palette color at 0 <= x < 256.

For instance, these functions can be used to sample a color like

    const uint16_t* palette = palettefx_get_palette_data_by_index(
        PALETTEFX_CARNIVAL);
    hsv_t color = palettefx_interp_color(palette, 64);