Select Word: improve multiline selection.
Thanks to @arkanoryn for catching this.

Previously when using line selection to select multiple lines, there was
a limitation where the selection might not extend to the end of the
line. As in @arkanoryn's example:

    abcd
    abcdefgh

"If I run "select_line" from line 1, and press it again, with the
current code I would select `abcd\nabcd`. the `efgh` would stay
unselected."

This commit fixes multiline line selection by sending a tap of Shift+End
(or on Mac, GUI+Shift+Right) when line select is released.