Speculative Hold: apply mod-tap mods immediately.
Speculative Hold is a QMK community module that modifies mod-tap `MT`
keys. The mod is applied immediately on key down, eliminating delay,
particularly when using mod-tap keys in tandem with an external mouse.
Depending on how the mod-tap key is settled, the hold action is canceled
(when the mod-tap is being tapped) or allowed to continue (when the
mod-tap is being held).

For full documentation, see
<https://getreuer.info/posts/keyboards/speculative-hold>